requirements.txt
google-api-python-client==2.153.0
google-auth==2.36.0
requests==2.32.3
instagrapi==2.1.2
